pyproject.toml
[project]
name = "medical-rag"
version = "0.1.0"
description = "Assist in diagnosing and treating medical conditions"
authors = [
    {name = "Muhammad Zaid",email = "zaidasif011@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "pinecone-text (>=0.10.0,<0.11.0)",
    "pinecone[grpc] (>=6.0.2,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "google-genai (>=1.7.0,<2.0.0)",
    "streamlit (>=1.43.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
